move documentation and interface to separate file
